# openapi.yaml
openapi: 3.1.0
info:
  title: EZVIZ VTM Bridge
  version: 1.7.0
  description: Internal bounded media contract for Home Assistant and SceneTrove.
servers: [{url: http://ezviz-vtm-bridge:8765}]
paths:
...
  /v1/cameras/{camera}/identity:
    get:
      summary: Resolve the physical serial/channel bound to a camera alias
      security: [{bearerAuth: []}, {basicAuth: []}]
      parameters: [{$ref: "#/components/parameters/Camera"}]
      responses: {"200": {description: Authenticated physical-camera binding, content: {application/json: {schema: {type: object, additionalProperties: false, required: [camera, source_id], properties: {camera: {type: string}, source_id: {type: string}}}}}}, "404": {$ref: "#/components/responses/NotFound"}}
  /v1/cameras/{camera}/snapshot.jpg:
    get:
      security: [{bearerAuth: []}, {basicAuth: []}]
      parameters: [{$ref: "#/components/parameters/Camera"}]
      responses:
        "200":
          description: Fresh or briefly coalesced JPEG snapshot
          headers:
            Cache-Control: {schema: {type: string, const: no-store}}
          content: {image/jpeg: {schema: {type: string, contentEncoding: binary}}}
        "404": {$ref: "#/components/responses/NotFound"}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer}
    basicAuth:
      type: http
      scheme: basic
      description: Fixed username homeassistant; token is the password.
  parameters:
    Camera: {name: camera, in: path, required: true, schema: {type: string, pattern: "^[A-Za-z0-9_-]+$"}}
...
  responses:
    NotFound: {description: Resource is absent or not ready}
security: [{bearerAuth: []}]